third_party/WebKit/LayoutTests/typedcssom/cssRotation-expected.txt

This is a testharness.js-based test.
FAIL Test that the (x, y, z, angle) values for CSSRotation are correct. assert_equals: expected (number) 0 but got (undefined) undefined
PASS Test that the is2D values for CSSRotation is correct. 
FAIL Test that toString values for CSSRotation is correct. assert_equals: expected "rotate(0deg)" but got ""
FAIL Test that asMatrix is constructed correctly for CSSRotation. Cannot read property 'is2D' of null
Harness: the test ran to completion.
